QRCode.jl Documentation

QRCode.jl Documentation

Module that can create QR codes as data or images using qrcode or exportqrcode.

Creating QR codes

QRCode.qrcode — Function.

qrcode(message::AbstractString, eclevel = Medium(); compact = false)

Create a BitArray{2} with the encoded message, with true (1) for the black areas and false (0) as the white ones. If compact is false, white space is added around the QR code.

The error correction level eclevel can be picked from four values: Low() (7% of missing data can be restored), Medium() (15%), Quartile() (25%) or High() (30%). Higher levels make denser QR codes.

source

QRCode.exportqrcode — Function.

exportqrcode( message::AbstractString
            , path = "qrcode.png"
            , eclevel = Medium()
            ; targetsize = 5
            , compact = false )

Create a PNG file with the encoded message of approximate size targetsize cm. If compact is false, white space is added around the QR code.

The error correction level eclevel can be picked from four values: Low() (7% of missing data can be restored), Medium() (15%), Quartile() (25%) or High() (30%). Higher levels make denser QR codes.

source

QRCode.getversion — Function.

getversion(message::AbstractString, mode::Mode, level::ErrCorrLevel)

Return the version of the QR code, between 1 and 40.

julia> getversion("Hello World!", Alphanumeric(), High())
2

source

QRCode.getmode — Function.

getmode(message::AbstractString)

Return the encoding mode of message, either Numeric(), Alphanumeric() or Byte().

Examples

julia> getmode("HELLO WORLD")
Alphanumeric()

source

Encoding modes

There are three several encoding mode currently supported.

QRCode.Mode — Type.

Abstract type that groups the three supported encoding modes Numeric, Alphanumeric and Byte.

source

QRCode.Numeric — Type.

Encoding mode for messages composed of digits only.

source

QRCode.Alphanumeric — Type.

Encoding mode for messages composed of digits, characters A-Z (capital only) , space and % * + - . / : $.

source

QRCode.Byte — Type.

Encoding mode for messages composed of ISO 8859-1 or UTF-8 characters.

source

Error Correction

There are four error correction levels you can choose from.

QRCode.ErrCorrLevel — Type.

Abstract type that groups the four error correction levels Low, Medium, Quartile and High.

source

QRCode.Low — Type.

Error correction level that can restor up to 7% of missing codewords.

source

QRCode.Medium — Type.

Error correction level that can restor up to 15% of missing codewords.

source

QRCode.Quartile — Type.

Error correction level that can restor up to 25% of missing codewords.

source

QRCode.High — Type.

Error correction level that can restor up to 30% of missing codewords.

source